// src/interfaces/Interfaces.tsx
// TODO: Create an interface for the Candidate objects returned by the API
export interface githubProfilesObject {
avatar_url: string
events_url:string
followers_url: string
following_url: string 
gists_url: string 
gravatar_id: string 
html_url: string
id: number
login: string 
node_id: string
organizations_url: string 
received_events_url: string 
repos_url: string
site_admin: boolean
starred_url: string
subscriptions_url: string
type: string
url: string
}

export interface GithubUser {
    avatar_url: string;
    bio: string | null;
    blog: string;
    company: string | null;
    created_at: string; // ISO date string
    email: string;
    events_url: string;
    followers: number;
    followers_url: string;
    following: number;
    following_url: string;
    gists_url: string;
    gravatar_id: string;
    hireable: boolean | null;
    html_url: string;
    id: number;
    location: string | null;
    login: string;
    name: string | null;
    node_id: string;
    organizations_url: string;
    public_gists: number;
    public_repos: number;
    received_events_url: string;
    repos_url: string;
    site_admin: boolean;
    starred_url: string;
    subscriptions_url: string;
    twitter_username: string | null;
    type: string;
    updated_at: string; // ISO date string
    url: string;
  }